+++ crates/module-system/module-implementations/sov-blob-storage/tests/integration/base_sequencing.rs
@@ -68,7 +68,7 @@ fn check_blob_selection() {
         );
 
         let result = runner.execute::<RelevantBlobs<MockBlob>>(slot_to_send);
-        assert_eq!(result.batch_receipts.len(), 3);
+        assert_eq!(result.0.batch_receipts.len(), 3);
     }
 
     {
@@ -82,7 +82,7 @@ fn check_blob_selection() {
         );
 
         let result = runner.execute::<RelevantBlobs<MockBlob>>(slot_to_send);
-        assert_eq!(result.batch_receipts.len(), 1);
+        assert_eq!(result.0.batch_receipts.len(), 1);
     }
 
     // Test the edge cases.
@@ -96,13 +96,13 @@ fn check_blob_selection() {
         );
 
         let result = runner.execute::<RelevantBlobs<MockBlob>>(slot_to_send);
-        assert_eq!(result.batch_receipts.len(), 1);
+        assert_eq!(result.0.batch_receipts.len(), 1);
     }
 
     {
         let slot_to_send = build_basic_blobs(&vec![], &mut nonces);
 
         let result = runner.execute::<RelevantBlobs<MockBlob>>(slot_to_send);
-        assert_eq!(result.batch_receipts.len(), 0);
+        assert_eq!(result.0.batch_receipts.len(), 0);
     }
 }