+++ crates/module-system/sov-test-utils/src/interface/tests.rs
@@ -7,6 +7,7 @@ pub use sov_modules_stf_blueprint::TxReceiptContents;
 use sov_state::{Storage, StorageProof};
 
 use super::{BatchType, ProofInput, TransactionType};
+use crate::runtime::BlobInfo;
 
 type TestAssertion<Context, S> = Box<dyn FnOnce(Context, &mut ApiStateAccessor<S>)>;
 type BatchReceipt<S> =
@@ -34,6 +35,8 @@ pub struct TransactionAssertContext<S: Spec, RT: RuntimeEventProcessor> {
     /// ```
     ///
     pub events: Vec<RT::RuntimeEvent>,
+    /// The metadata about the blob that contained the transaction
+    pub blob_info: BlobInfo,
     /// The outcome of the transaction.
     pub tx_receipt: TxEffect<TxReceiptContents<S>>,
 }
@@ -42,6 +45,7 @@ impl<S: Spec, RT: RuntimeEventProcessor> TransactionAssertContext<S, RT> {
     /// Creates a [`TransactionAssertContext`] from the given [`TransactionReceipt`].
     pub fn from_receipt<Da: DaSpec>(
         receipt: TransactionReceipt<TxReceiptContents<S>>,
+        blob_info: BlobInfo,
         gas_value_used: u64,
     ) -> Self {
         let events = receipt
@@ -57,6 +61,7 @@ impl<S: Spec, RT: RuntimeEventProcessor> TransactionAssertContext<S, RT> {
         TransactionAssertContext {
             tx_receipt: receipt.receipt,
             events,
+            blob_info,
             gas_value_used,
         }
     }